updates models and controllers to split /api and /api/topics into seperate files
