+++ controllers/topics.controllers.js
@@ -1,17 +1,9 @@
-const { fetchTopics, fetchEndpoints } = require("../models/topics.models.js")
+const { fetchTopics } = require("../models/topics.models.js")
 
 exports.getTopics = (req, res, next) => {
     fetchTopics()
         .then((topics) => {
             res.status(200).send(topics)
-        })
-        .catch((err) => next(err));
-}
-
-exports.getEndpoints = (req, res, next) => {
-    fetchEndpoints()
-        .then((endpoints) => {
-            res.status(200).send(endpoints)
         })
         .catch((err) => next(err));
 }
