+++ models/topics.models.js
@@ -6,10 +6,3 @@ exports.fetchTopics = () => {
         .query("SELECT * FROM topics;")
         .then(({rows}) => rows)
 }
-
-exports.fetchEndpoints = () => {
-    return fs.readFile(`${__dirname}/../endpoints.json`, {encoding: "utf-8"}).then((endpointsData) => {
-        const parsedEndpointsData = JSON.parse(endpointsData);
-        return parsedEndpointsData
-    })
-}
